+++ weblate/trans/models/translation.py
@@ -21,7 +21,7 @@ import codecs
 import os
 import tempfile
 from datetime import datetime
-from typing import BinaryIO, Optional
+from typing import BinaryIO, Dict, Optional
 
 from django.core.cache import cache
 from django.core.exceptions import ValidationError
@@ -1045,19 +1045,39 @@ class Translation(models.Model, URLMixin, LoggerMixin, CacheKeyMixin):
             author=user,
         )
 
-    def new_unit(self, request, key, value):
+    def new_unit(
+        self,
+        request,
+        key: Optional[str],
+        value: Optional[str],
+        batch: Optional[Dict[str, str]] = None,
+    ):
+        from weblate.auth.models import get_anonymous
+
+        user = request.user if request else get_anonymous()
         with self.component.repository.lock:
-            self.component.commit_pending("new unit", request.user)
-            Change.objects.create(
-                translation=self,
-                action=Change.ACTION_NEW_UNIT,
-                target=value,
-                user=request.user,
-                author=request.user,
-            )
-            self.store.new_unit(key, value)
+            self.component.commit_pending("new unit", user)
+            if batch:
+                for key, value in batch.items():
+                    self.store.new_unit(key, value)
+                    Change.objects.create(
+                        translation=self,
+                        action=Change.ACTION_NEW_UNIT,
+                        target=value,
+                        user=user,
+                        author=user,
+                    )
+            else:
+                self.store.new_unit(key, value)
+                Change.objects.create(
+                    translation=self,
+                    action=Change.ACTION_NEW_UNIT,
+                    target=value,
+                    user=user,
+                    author=user,
+                )
             self.component.create_translations(request=request)
-            self.git_commit(request.user, request.user.get_author_name())
+            self.git_commit(user, user.get_author_name())
 
 
 class GhostTranslation: